fin de lupload

+++ back/essai.py
@@ -67,8 +67,8 @@ def LaMuse():
     return "Finish"
 
 
-@app.route('/uploadFile', methods=['GET', 'POST'])
-def upload_file():
+@app.route('/uploadBack', methods=['GET', 'POST'])
+def upload_Back():
     if request.method == 'POST':
         # check if the post request has the file part
         if 'file' not in request.files:
@@ -83,13 +83,38 @@ def upload_file():
         if file and allowed_file(file.filename) :
             filename = secure_filename(file.filename)
             os.system("rm -f "+back_path+'*')
-            os.system("rm -f "+paint_path+'*')
             os.system("rm -f "+result_path+'*')
             
             app.config['UPLOAD_FOLDER'] = back_path
             file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
             resp = jsonify(success=True)
             resp.status_code = 200
+            return resp
+
+@app.route('/uploadPaint', methods=['GET', 'POST'])
+def upload_Paint():
+    if request.method == 'POST':
+        print('UploadPaint')
+        # check if the post request has the file part
+        if 'file' not in request.files:
+            flash('No file part')
+            return redirect(request.url)
+        file = request.files['file']
+        # if user does not select file, browser also
+        # submit an empty part without filename
+        if file.filename == '':
+            flash('No selected file')
+            return redirect(request.url)
+        if file and allowed_file(file.filename) :
+            filename = secure_filename(file.filename)
+            os.system("rm -f "+paint_path+'*')
+            os.system("rm -f "+result_path+'*')
+            
+            app.config['UPLOAD_FOLDER'] = paint_path
+            file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
+            resp = jsonify(success=True)
+            resp.status_code = 200
+            
             return resp
